Clean: Removed the deprecated store and changed the name of the hex viewer app
